riman_fashion_erp/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-environ==0.12.0
python-decouple==3.8
PyJWT==2.8.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
reportlab==4.0.9
Pillow==10.1.0
xlsxwriter==3.1.8
openpyxl==3.11.0
